Rename web module to wasm
To more accurately reflect that the module is used to output wasm -- not
necessarily run on the web (it can run in node js just fine)
