+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "src/common",
-    "src/web",
+    "src/wasm",
     "src/native",
 ]
