+++ package.json
@@ -12,17 +12,18 @@
     "./dist/node/cjs/index_node.js": "./dist/browser/cjs/index_browser.js",
     "./dist/node/es/index_node.js": "./dist/browser/es/index_browser.js"
   },
-  "types": "./dist/browser/es/index_browser.d.ts",
+  "types": "./dist/browser/es/main/index_browser.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist/ && npm run build:web && npm run build:web-simd && npm run build:native && npm run build:bundle",
+    "build": "rm -rf dist/ && npm run build:wasm && npm run build:wasm-simd && npm run build:native && npm run build:bundle",
     "build:native": "bash ./assets/build-native.sh",
-    "build:web": "wasm-pack build -t web --out-dir ../main/web src/web",
-    "build:web-simd": "RUSTFLAGS='-C target-feature=+simd128' wasm-pack build -t web --out-dir ../main/web-simd src/web",
+    "build:wasm": "wasm-pack build -t web --out-dir ../main/wasm src/wasm",
+    "build:wasm-simd": "RUSTFLAGS='-C target-feature=+simd128' npm run build:wasm-simd-ci",
+    "build:wasm-simd-ci": "wasm-pack build -t web --out-dir ../main/wasm-simd src/wasm",
     "build:bundle": "rollup -c",
-    "optimize": "wasm-opt -Os src/main/web/highwayhasher_web_bg.wasm -o tmp.wasm && mv tmp.wasm src/main/web/highwayhasher_web_bg.wasm && wasm-opt -Os src/main/web-simd/highwayhasher_web_bg.wasm -o tmp.wasm && mv tmp.wasm src/main/web-simd/highwayhasher_web_bg.wasm",
+    "optimize": "wasm-opt -Os src/main/wasm/highwayhasher_web_bg.wasm -o tmp.wasm && mv tmp.wasm src/main/wasm/highwayhasher_web_bg.wasm && wasm-opt -Os src/main/wasm-simd/highwayhasher_web_bg.wasm -o tmp.wasm && mv tmp.wasm src/main/wasm-simd/highwayhasher_web_bg.wasm",
     "test": "npm run build && npm run test:inplace",
     "test:inplace": "npm run test:native && npm run test:types",
     "test:types": "tsc --noEmit tests/unit/usage.ts",
