+++ src/main/index_browser.ts
@@ -1,3 +1,3 @@
 export type { IHash, HashCreator, HighwayLoadOptions } from "./model";
-export { WasmHighwayHash, hasSimd as hasWasmSimd } from "./web";
-export { WasmHighwayHash as HighwayHash } from "./web";
+export { WasmHighwayHash, hasSimd as hasWasmSimd } from "./wasm";
+export { WasmHighwayHash as HighwayHash } from "./wasm";
