+++ src/main/index_node.ts
@@ -1,3 +1,3 @@
 export type { IHash, HashCreator, HighwayLoadOptions } from "./model";
-export { WasmHighwayHash, hasSimd as hasWasmSimd } from "./web";
+export { WasmHighwayHash, hasSimd as hasWasmSimd } from "./wasm";
 export { NativeHighwayHash as HighwayHash } from "./native";
