+++ src/wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "highwayhasher-web"
+name = "highwayhasher-wasm"
 version = "0.1.0"
 authors = ["nickbabcock <nbabcock19@hotmail.com>"]
 edition = "2018"
